refactor: organize project into clean architecture structure
