+++ internal/domain/entities/response.go
